+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`-` New email placeholder: ``{all_reviews}`` allows you to send all review texts (though not scores!) to submitters.
 - :bug:`1266` pretalx only recognised overlapping scheduled talks for a speaker when they didn't start or end at the exact same time.
 - :feature:`-` The rendering speed of all backend pages has been improved.
 - :feature:`-` The performance of the schedule editor and release pages was improved for large events.